managed-connectivity/community-contributed-connectors/aws-glue-connector/requirements.txt
google-cloud-dataplex>=2.4.0
boto3
google-cloud-secret-manager
google-cloud-storage